Remove client wrapper
